Fix (othersPage.jsx: othersPage 경로 부활에 따른 모듈 경로 수정), Rename (src/components/features/ohters/UserProfileContainer.jsx -> src/components/features/ohterPage/UserProfileContainer.jsx)
