+++ src/pages/othersPage.jsx
@@ -2,7 +2,7 @@
 import Layout from "../layout/layout";
 import Footer from "../components/common/footer";
 import Header from "../components/common/header";
-import UserProfileContainer from "../components/features/ohters/UserProfileContainer";
+import UserProfileContainer from "../components/features/ohterPage/UserProfileContainer";
 
 function OthersPage() {
   return (
